initial pdf annotation retrieval
added call to get file mimetype and switch between docx and pdf
added streaming processor class

pdf differs in that there aren't anchor events in the text, but one can
instead get the anchor text. Where that text is in the overall document
stream has to be figured out.

The retrieval of annotations only looks for subtype 'highlight' and
currently only looks for them on page 1.